# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhooks": {
      "get": {
        "operationId": "WebhookController_getWebhooks",
        "summary": "Get Webhooks",
        "parameters": [
          {
            "name": "field",
            "required": false,
            "in": "query",
            "example": "_id",
            "description": "field that you want to find",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "values",
            "required": false,
            "in": "query",
            "example": "[\"6200e7b2b3cfeca3405f57f0\", \"61fd2eecf3f78493727204f0\"]",
            "description": "JSON array with values that you want to find",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "example": 1649186190881,
            "description": "UTC time from (created_at)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "example": 1649186190881,
            "description": "UTC time to (created_at)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "example": 100,
            "description": "limit number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "example": 5,
            "description": "skip number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebHooks"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Webhook API"
        ]
      },
      "post": {
        "operationId": "WebhookController_createNewWebhook",
        "summary": "Create new Webhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebHooks"
                }
              }
            }
          }
        },
        "tags": [
          "Webhook API"
        ]
      }
    },
    "/webhooks/{webhookId}": {
      "put": {
        "operationId": "WebhookController_updateWebhook",
        "summary": "Update Webhook",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "example": "6200e7b2b3cfeca3405f57f0",
            "description": "webhook history id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebHooks"
                }
              }
            }
          }
        },
        "tags": [
          "Webhook API"
        ]
      },
      "delete": {
        "operationId": "WebhookController_deleteWebhook",
        "summary": "Delete Webhook",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "example": "6200e7b2b3cfeca3405f57f0",
            "description": "webhook history id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/String"
                }
              }
            }
          }
        },
        "tags": [
          "Webhook API"
        ]
      }
    },
...
  },
  "info": {
    "title": "System for optimization of maprouting",
    "description": "Documentation REST API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "String": {
        "type": "object",
        "properties": {}
      },
...
      "WebHooks": {
        "type": "object",
        "properties": {}
      },
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://your.domain/webhooks",
            "description": "POST url to server that listen webhooks"
          },
          "event": {
            "type": "string",
            "example": "Error Created",
            "description": "webhook name"
          },
          "name": {
            "type": "string",
            "example": "error_created",
            "description": "your webhook name"
          }
        },
        "required": [
          "url",
          "event",
          "name"
        ]
      },
      "UpdateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://your.domain/webhooks",
            "description": "POST url to server that listen webhooks"
          },
          "event": {
            "type": "string",
            "example": "Error Created",
            "description": "webhook name"
          },
          "name": {
            "type": "string",
            "example": "error_created",
            "description": "your webhook name"
          }
        }
      },
...
    }
  }
}